+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixes
+
+- Update AADSyncRule class to exclude all properties when comparing states for
+  standard rules except `Name` and `Disabled`. This is because standard rules
+  cannot be changed. A second compare job is executed for reporting on
+  differences but without having an effect on the overall test result.
+
+
 ## [0.3.2] - 2025-07-22
 
 ### Changed